Reconstruct shop list animation, inventory queries and background transfers

+++ include/game/overlay005_display.h
@@ -105,7 +105,7 @@ void Overlay5Display_LoadObjTiles(
     DisplayEngine engine, const void *source, u32 offset, u32 size
 );
 
-void Overlay5DisplayBg_SetVerticalOffset(DisplayEngine engine, int background, s16 y);
+void Overlay5DisplayBg_SetVerticalOffset(DisplayEngine engine, int background, int y);
 void Overlay5DisplayBg_SetOffset(DisplayEngine engine, int background, s16 x, s16 y);
 void Overlay5DisplayBg_SetOverflow(DisplayEngine engine, int background, u16 enabled);
 void Overlay5DisplayBg_SetAreaOverflow(DisplayEngine engine, int background, u16 enabled);
